이미지 crud

+++ src/main/java/com/uts3back/controller/ImagesController.java
@@ -5,6 +5,7 @@ import com.uts3back.service.ImagesService;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,6 +24,7 @@ public class ImagesController {
             @RequestPart("file") MultipartFile file,
             @RequestPart("user-service-id") String userServiceID) throws Exception {
 
+
         imagesService.uploadFileService(file, userServiceID);
         return ResponseEntity.ok("이미지 업로드 성공");
     }
@@ -32,6 +34,13 @@ public class ImagesController {
     public ResponseEntity<String> updateImage(
             @PathVariable("img-id") String imgID,
             @RequestPart("file") MultipartFile file) {
+
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        // 요청한 사용자가 이미지의 소유자인지 확인합니다.
+        if (!imagesService.findEmailByImgID(imgID).equals(email)) {
+            return ResponseEntity.status(403).body("이미지 수정 권한이 없습니다.");
+        }
+        
         imagesService.updateImage(imgID, file);
         return ResponseEntity.ok("이미지 수정 완료");
     }
@@ -39,17 +48,32 @@ public class ImagesController {
     @Operation(summary = "이미지 삭제", description = "이미지 삭제")
     @DeleteMapping("/{img-id}")
     public ResponseEntity<String> deleteImage(@PathVariable("img-id") String imgID) {
+
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        // 요청한 사용자가 이미지의 소유자인지 확인합니다.
+        if (!imagesService.findEmailByImgID(imgID).equals(email)) {
+            return ResponseEntity.status(403).body("이미지 삭제 권한이 없습니다.");
+        }
+
         if(imagesService.deleteImage(imgID)){
             return ResponseEntity.ok("이미지 삭제 완료");
         }else{
-            return ResponseEntity.status(404).body(null);
+            return ResponseEntity.status(404).body("이미지가 존재하지 않음");
         }
     }
 
     // 조회 컨트롤러 추가
     @Operation(summary = "이미지 조회", description = "이미지 정보 조회")
     @GetMapping("/{img-id}")
-    public ResponseEntity<ImagesDTO> getImage(@PathVariable("img-id") String imgID) {
+    public ResponseEntity getImage(@PathVariable("img-id") String imgID) {
+
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        // 요청한 사용자가 이미지의 소유자인지 확인합니다.
+        if (!imagesService.findEmailByImgID(imgID).equals(email)) {
+            return ResponseEntity.status(403).body("이미지 조회 권한이 없습니다.");
+        }
+
         ImagesDTO imagesDTO = imagesService.getImage(imgID);
         return ResponseEntity.ok(imagesDTO);
     }
